requirements.txt
torch==2.11.0
torchvision==0.26.0
numpy==2.4.2
Pillow==12.1.0
lmdb>=1.4.0
higher>=0.2.1
matplotlib>=3.7.0
pandas==3.0.0
seaborn>=0.12.0
wandb==0.18.0
efficientnet-pytorch>=0.7.1
PyYAML>=6.0
